feature: category 수정 및 삭제 기능 예외처리

+++ src/main/java/com/pm/projectmanager/domain/category/CategoryService.java
@@ -3,9 +3,11 @@ package com.pm.projectmanager.domain.category;
 import com.pm.projectmanager.common.response.ResponseExceptionEnum;
 import com.pm.projectmanager.domain.authority.Authority;
 import com.pm.projectmanager.domain.authority.AuthorityRepository;
+import com.pm.projectmanager.domain.card.CardRepository;
 import com.pm.projectmanager.domain.category.dto.*;
 import com.pm.projectmanager.domain.user.User;
 import com.pm.projectmanager.exception.AuthorityNullException;
+import com.pm.projectmanager.exception.CannotDeleteCategoryWithRelatedCardsException;
 import com.pm.projectmanager.exception.CategoryNameAlreadyExistsInProjectException;
 import com.pm.projectmanager.exception.CategoryNotFoundException;
 import jakarta.transaction.Transactional;
@@ -13,7 +15,9 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
+import java.sql.SQLIntegrityConstraintViolationException;
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -21,6 +25,7 @@ public class CategoryService {
 
     private final CategoryRepository categoryRepository;
     private final AuthorityRepository authorityRepository;
+    private final CardRepository cardRepository;
 
     // 카테고리 생성
     public void createCategory(CreateCategoryRequestDto requestDto, User user, Long projectId) {
@@ -47,6 +52,12 @@ public class CategoryService {
     @Transactional
     public void updateCategory(UpdateCategoryRequestDto requestDto, User user, Long categoryId, Long projectId) {
         hasProjectAndUser(projectId, user.getId());
+
+        Optional<Category> existingCategory = categoryRepository.findByProjectIdAndName(projectId, requestDto.getName());
+        if(existingCategory.isPresent()) {
+            throw new CategoryNameAlreadyExistsInProjectException(
+                    ResponseExceptionEnum.CATEGORY_DELETE_RELATED_CARDS_EXIST);
+        }
         Category category = categoryRepository.findById(categoryId).orElseThrow(
                 () -> new CategoryNotFoundException(ResponseExceptionEnum.CATEGORY_NOT_FOUND));
         category.update(requestDto.getColor(), requestDto.getName(), requestDto.getDescription());
@@ -57,6 +68,10 @@ public class CategoryService {
         hasProjectAndUser(projectId, user.getId());
         Category category = categoryRepository.findById(categoryId).orElseThrow(
                 () -> new CategoryNotFoundException(ResponseExceptionEnum.CATEGORY_NOT_FOUND));
+        if(cardRepository.existsByCategoryId(categoryId)) {
+            throw new CannotDeleteCategoryWithRelatedCardsException(
+                    ResponseExceptionEnum.CATEGORY_DELETE_RELATED_CARDS_EXIST);
+        }
         categoryRepository.delete(category);
     }
